+++ test/Version.t.sol
@@ -8,6 +8,7 @@ import {RuleBlacklist} from "src/rules/validation/deployment/RuleBlacklist.sol";
 import {RuleSanctionsList} from "src/rules/validation/deployment/RuleSanctionsList.sol";
 import {ISanctionsList} from "src/rules/interfaces/ISanctionsList.sol";
 import {RuleMaxTotalSupply} from "src/rules/validation/deployment/RuleMaxTotalSupply.sol";
+import {TotalSupplyMock} from "src/mocks/TotalSupplyMock.sol";
 import {RuleWhitelistWrapper} from "src/rules/validation/deployment/RuleWhitelistWrapper.sol";
 import {RuleERC2980} from "src/rules/validation/deployment/RuleERC2980.sol";
 import {RuleConditionalTransferLight} from "src/rules/operation/RuleConditionalTransferLight.sol";
@@ -32,7 +33,7 @@ contract VersionTest is Test, HelperContract {
     }
 
     function testVersionRuleMaxTotalSupply() public {
-        RuleMaxTotalSupply rule = new RuleMaxTotalSupply(DEFAULT_ADMIN_ADDRESS, ADDRESS1, 0);
+        RuleMaxTotalSupply rule = new RuleMaxTotalSupply(DEFAULT_ADMIN_ADDRESS, address(new TotalSupplyMock()), 0);
         assertEq(rule.version(), EXPECTED_VERSION);
     }